requirements.txt
streamlit>=1.36
supabase>=2.6.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26
openpyxl>=3.1.5
XlsxWriter>=3.2.0
reportlab>=4.2.0